+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [2.0.0] - 2019-06-21
+## [2.0.1] - 2019-06-21
 ### Added
 - Dropdown placeholder text + language key
 - Dropdown appearance options: smart and smart_compact